requirements.txt
Django>=3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
tensorflow==2.3.1
Keras==2.4.3
python-dotenv==0.15.0
pymongo==3.11.2
bcrypt==3.2.0
dnspython==2.0.0
# requirements-dev
sphinx-markdown-tables==0.0.15
pydata-sphinx-theme==0.5.0
recommonmark==0.7.1
